+++ vite.config.ts
@@ -10,6 +10,7 @@ export default defineConfig({
     react(),
     legacy()
   ],
+  base:"/it35-lab"
   test: {
     globals: true,
     environment: 'jsdom',
